tests: move dfx-mgr platform selection to integration tests
due to needing available to return true for dfx-mgr softener
